fix comments (continue)
