+++ HomeEdu/Support/Constant.swift
@@ -31,9 +31,6 @@ enum Constant {
     static let schedule = "new_schedule"
     static let exam = "new_examschedule"
     
-    //General Information
-    static let studentId = "16022411"
-    
     //Errors
     static let dataErr = "Data don't have any items"
 }
