+++ HomeEdu/Support/StudentTest.swift
@@ -1,0 +1,11 @@
+//
+//  StudentTest.swift
+//  HomeEdu
+//
+//  Created by Minh Mon on 8/1/19.
+//  Copyright © 2019 Nguyen Cuong. All rights reserved.
+//
+
+import Foundation
+
+var studentId = "16022411"
